feat: Prelim behavior consolidation
