+++ src/config.ts
@@ -1,5 +1,3 @@
-import path from "path";
-import os from "os";
 // eslint-disable-next-line no-restricted-imports
 import type fs from "fs";
 import { z } from "zod";
@@ -10,13 +8,7 @@ import {
   Settings,
   StateData,
 } from "./types";
-
-export const APPDATA_DIR = path.join(os.homedir(), ".code-container");
-export const CONFIGS_DIR = path.join(APPDATA_DIR, "configs");
-export const TEMP_DIR = path.join(APPDATA_DIR, "temp");
-export const SETTINGS_PATH = path.join(APPDATA_DIR, "settings.json");
-export const STATE_PATH = path.join(TEMP_DIR, "state.json");
-export const USER_DOCKERFILE_PATH = path.join(APPDATA_DIR, "Dockerfile.User");
+import { APPDATA_DIR, CONFIGS_DIR, TEMP_DIR } from "./platform/paths";
 
 export type FsReader = Pick<
   typeof fs,
