+++ src/platform/os.ts
@@ -1,0 +1,24 @@
+export enum Platform {
+  Windows = "win32",
+  Linux = "linux",
+  Macos = "darwin",
+}
+
+export function getPlatform(): Platform {
+  const p = process.platform;
+  if (p === "win32") return Platform.Windows;
+  if (p === "darwin") return Platform.Macos;
+  return Platform.Linux;
+}
+
+export function isWindows(): boolean {
+  return getPlatform() === Platform.Windows;
+}
+
+export function isLinux(): boolean {
+  return getPlatform() === Platform.Linux;
+}
+
+export function isMacos(): boolean {
+  return getPlatform() === Platform.Macos;
+}
